Refactor: 인증 관련 파일 이름 재정의
